setup mixpanel proxy
