+++ docker-compose.yaml
@@ -1,4 +1,3 @@
-version: "0.1"
 services:
   api:
     build: .
